+++ CHANGELOG.md
@@ -6,8 +6,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Install kubectl on k3d setup automatically; #71
+- Use workspace as k3d constructor parameter
 
-## [1.49.0](https://github.com/cloudogu/ces-build-lib/releases/tag/1.50.0) - 2022-03-10
+### Removed
+- Remove unused git credentials parameter in k3d constructor
+
+## [1.50.0](https://github.com/cloudogu/ces-build-lib/releases/tag/1.50.0) - 2022-03-10
 ### Changed
 - Install Gitops Playground and K3d into a subfolder for better CI experience
   - see the README.md for updated usage tips